+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 
 # Changelog
 
+## 4.1.11 (15/MAR/2022)
+* Fixes issue where bitrate is not reported for streams without textTracks.
+
 ## 4.1.10 (30/SEP/2021)
 * Supports Video.js 7.15.2.
-* Fixes issue bitrate is not reported for HLS demux stream.
+* Fixes issue where bitrate is not reported for HLS demux stream.
 * Fixes issue of error when dropped video frames api is not available.
 
 ## 4.1.8 (30/MAR/2021)